Game_Engine/Chess/requirements.txt
pygame==2.1.2
pygame-menu==4.2.2
Pillow==8.0.0